requirements.txt
pillow>=9.0.0
pystray>=0.19.0
psutil>=5.9.0
requests>=2.28.0